Adding separate id for each set of tests + color console to differenciate the tests + removing end test as a test

+++ src/module/extension/webserver/CatObjects/assert.js
@@ -1,17 +1,9 @@
 var _jmr = require("test-model-reporter"),
     _global = catrequire("cat.global"),
     _log = _global.log(),
-    _reportCreator,
-    _testsuite = _jmr.create({
-        type: "model.testsuite",
-        data: {
-            name: "testsuite"
-        }
-    }),
+    _reportCreator = {},
     _catcli = (catrequire ? catrequire("cat.cli") : null),
     _fs = require("fs"),
-    _testConfigMap,
-    _isAlive = false, //guilty until proven innocent
     _checkIfAlive; //TODO: make this configurable
 
 function readConfig() {
@@ -30,7 +22,7 @@ function readConfig() {
             try {
                 sourceFolder = project.getInfo("source");
                 configPath = path.join(sourceFolder, "/config/cat.json");
-            } catch(e) {
+            } catch (e) {
                 _log.error("[CAT server (assert module)] Failed to load cat.json test project, No CAT test project is available.", e);
             }
         } else {
@@ -50,11 +42,22 @@ function readConfig() {
     return testConfigMap;
 }
 
-
-_testConfigMap = readConfig();
-
 function ReportCreator(filename) {
     this._fileName = filename;
+    this._testConfigMap = readConfig();
+    this._hasFailed = false;
+    this._testsuite = _jmr.create({
+        type: "model.testsuite",
+        data: {
+            name: "testsuite"
+        }
+    });
+
+    var colorsArray = ['grey', 'cyan', 'grey', 'grey', 'cyan', 'yellow', 'blue'];
+    this._randomColor = colorsArray[Math.floor(Math.random() * colorsArray.length)];
+}
+ReportCreator.prototype.getTestConfigMap = function () {
+    return this._testConfigMap;
 }
 
 ReportCreator.prototype.addTestCase = function (testName, status, phantomStatus, message) {
@@ -81,44 +84,39 @@ ReportCreator.prototype.addTestCase = function (testName, status, phantomStatus,
         testCase.add(result);
     }
 
-    if (_testConfigMap[testName]) {
-        _testConfigMap[testName].wasRun = true;
+    if (this._testConfigMap[testName]) {
+        this._testConfigMap[testName].wasRun = true;
     }
     else {
         _log.info("Test " + testName + " not in test manager");
     }
 
-    _testsuite.add(testCase);
+    if (status !== 'End') {
+        this._testsuite.add(testCase);
 
-    var output = _testsuite.compile();
+        var output = this._testsuite.compile();
 
-    var symbol = status === 'failure' ? '✖' : '✓',
-        statusDesc = status === 'failure' ? 'failed' : 'passed';
+        var symbol = status === 'failure' ? '✖' : '✓';
+        if (status === 'failure') {
+            this._hasFailed = true;
+        }
 
-    console.log(symbol + "Test " + testName + " " + message + " " + statusDesc);
+        var colors = require('colors');
+        colors.setTheme({'current' : this._randomColor});
 
-    if (_fs.existsSync(this._fileName)) {
-        _fs.unlinkSync(this._fileName);
+        var logmessage = symbol + "Test " + testName + " " + message;
+        console.log(logmessage.current);
+
+        if (_fs.existsSync(this._fileName)) {
+            _fs.unlinkSync(this._fileName);
+        }
+        _jmr.write(this._fileName, output);
+    } else {
+        var result = this._hasFailed ? "failed" : "succeeded";
+        console.log("======== Test End " + result + "========");
     }
-    _jmr.write(this._fileName, output);
 };
 
-_checkIfAlive = setInterval(function () {
-    if (!_isAlive) {
-        clearInterval(_checkIfAlive);
-        _log.info("Tests stopped reporting, probably a network problem, failing the rest of the tests");
-        if (!_reportCreator) {
-            _reportCreator = new ReportCreator("notestname.xml");
-        }
-
-        for (var key in _testConfigMap) {
-            _log.info(_testConfigMap[key]);
-            if (!_testConfigMap[key].wasRun) {
-                _reportCreator.addTestCase(_testConfigMap[key].name, 'failure', 'unknown', 'failed due to network issue');
-            }
-        }
-    }
-}, 30000); //TODO: make this configurable
 
 if (_jmr === undefined) {
     _log.info("Test Unit Reporter is not supported, consider adding it to the .catproject dependencies");
@@ -132,25 +130,46 @@ exports.result = function (req, res) {
         status = req.query.status,
         reportType = req.query.type,
         hasPhantom = req.query.hasPhantom,
+        id = req.query.id,
         file;
 
     /*  if (status === 'end') {
      clearInterval(_checkIfAlive);
      } else {*/
-    _isAlive = true;
+    clearTimeout(_checkIfAlive);
+    if (status !== 'End') {
+        _checkIfAlive = setTimeout(function () {
+            _log.info("Tests stopped reporting, probably a network problem, failing the rest of the tests");
+            if (_reportCreator == {}) {
+                _reportCreator['notest'] = new ReportCreator("notestname.xml");
+            }
+
+            for (var reportKey in _reportCreator) {
+                var testConfigMap = _reportCreator[reportKey].getTestConfigMap();
+
+                for (var key in testConfigMap) {
+                    _log.info(testConfigMap[key]);
+                    if (!testConfigMap[key].wasRun) {
+                        _reportCreator[reportKey].addTestCase(testConfigMap[key].name, 'failure', '', 'failed due to network issue');
+                    }
+                }
+            }
+
+        }, 30000); //TODO: make this configurable
+    }
     _log.info("requesting " + testName + message + status);
     res.setHeader('Content-Type', 'text/javascript;charset=UTF-8');
     res.send({"testName": testName, "message": message, "status": status});
 
 
     var phantomStatus = (hasPhantom === "true" ? "phantom" : "");
-    file = "./cattestresult" + reportType + "-" + phantomStatus + ".xml";
+    file = "./" + reportType + "-" + phantomStatus + id + ".xml";
 
-    if (!_reportCreator) {
-        _reportCreator = new ReportCreator(file);
+    if (!_reportCreator[id]) {
+        _reportCreator[id] = new ReportCreator(file);
     }
 
-    _reportCreator.addTestCase(testName, status, phantomStatus, message);
+    _reportCreator[id].addTestCase(testName, status, phantomStatus, message);
     // }
 };
